Original change: https://android-review.googlesource.com/c/platform/hardware/google/apf/+/2862612

+++ v5/test_buf_allocator.c
@@ -21,7 +21,7 @@
 
 uint8_t apf_test_buffer[APF_TX_BUFFER_SIZE];
 uint8_t apf_test_tx_packet[APF_TX_BUFFER_SIZE];
-uint32_t apf_test_tx_packet_len;
+int apf_test_tx_packet_len;
 uint8_t apf_test_tx_dscp;
 
 /**
@@ -46,6 +46,6 @@ int apf_transmit_buffer(__attribute__((unused)) void* ctx, uint8_t* ptr,
                         int len, uint8_t dscp) {
   apf_test_tx_packet_len = len;
   apf_test_tx_dscp = dscp;
-  memcpy(apf_test_tx_packet, ptr, len);
+  memcpy(apf_test_tx_packet, ptr, (size_t) len);
   return 0;
 }
